Update Build.PL in preparation for upgrading ext/xxHash to the current version.
The upstream xxHash repository now contains various C test files, in a
subdirectory. Unfortunately, the only API Module::Build offers by default
is "recursively scan the tree for all C files and compile and link the lot"

This doesn't work when some of the C files are test programs, and each
defines main(), meaning that the link stage files with an error about
duplicate symbols.

As Module::Build doesn't offer an option to filter, the least worst choice
is to subclass it and override process_support_files() to do what we need.
(Approach suggested by the current maintainer of Module::Build)